+++ src/font.rs
@@ -26,12 +26,22 @@ impl Font {
         }
     }
 
-    fn set(&self, name: String) {
+    fn set(&self, name: String, weight: font::Weight) {
         let name = Box::leak(name.into_boxed_str());
         let weight = if self.bold {
-            font::Weight::Bold
+            match weight {
+                font::Weight::Thin => font::Weight::Normal,
+                font::Weight::ExtraLight => font::Weight::Medium,
+                font::Weight::Light => font::Weight::Semibold,
+                font::Weight::Normal => font::Weight::Bold,
+                font::Weight::Medium => font::Weight::ExtraBold,
+                font::Weight::Semibold
+                | font::Weight::Bold
+                | font::Weight::ExtraBold
+                | font::Weight::Black => font::Weight::Black,
+            }
         } else {
-            font::Weight::Normal
+            weight
         };
         let style = if self.italics {
             font::Style::Italic
@@ -57,11 +67,13 @@ pub fn set(config: Option<&Config>) {
     let family = config
         .and_then(|config| config.font.family.clone())
         .unwrap_or_else(|| String::from("Iosevka Term"));
+    let weight =
+        config.map_or(font::Weight::Normal, |config| config.font.weight);
 
-    MONO.set(family.clone());
-    MONO_BOLD.set(family.clone());
-    MONO_ITALICS.set(family.clone());
-    MONO_BOLD_ITALICS.set(family);
+    MONO.set(family.clone(), weight);
+    MONO_BOLD.set(family.clone(), weight);
+    MONO_ITALICS.set(family.clone(), weight);
+    MONO_BOLD_ITALICS.set(family, weight);
 }
 
 pub fn load() -> Vec<Cow<'static, [u8]>> {